Reorganized folder structure with activities
