+++ app/src/main/java/com/canteen/app/activity/client/cart/OrderCartActivity.java
@@ -1,4 +1,4 @@
-package com.canteen.app.activity.cart;
+package com.canteen.app.activity.client.cart;
 
 import android.content.Context;
 import android.os.Bundle;
