+++ app/src/main/java/com/canteen/app/activity/client/cart/OrderItemsListViewHolder.java
@@ -1,4 +1,4 @@
-package com.canteen.app.activity.cart;
+package com.canteen.app.activity.client.cart;
 
 import android.support.constraint.ConstraintLayout;
 import android.support.v7.widget.RecyclerView;
