+++ app/src/main/java/com/canteen/app/activity/client/details/PriceContainer.java
@@ -1,4 +1,4 @@
-package com.canteen.app.activity.food.details;
+package com.canteen.app.activity.client.details;
 
 interface PriceContainer {
     void updatePrice(double priceIncrease);
