+++ app/src/main/java/com/canteen/app/activity/client/menu/details/MenuDetailsViewHolder.java
@@ -1,4 +1,4 @@
-package com.canteen.app.activity.menu.details;
+package com.canteen.app.activity.client.menu.details;
 
 import android.content.Intent;
 import android.support.constraint.ConstraintLayout;
@@ -8,7 +8,7 @@ import android.widget.Toast;
 
 import com.canteen.app.App;
 import com.canteen.app.R;
-import com.canteen.app.activity.food.details.FoodDetailsActivity;
+import com.canteen.app.activity.client.details.FoodDetailsActivity;
 import com.canteen.app.models.Food;
 import com.canteen.app.service.order.OrderCartService;
 import com.canteen.app.service.order.OrderItem;
