+++ app/src/main/res/layout/activity_add_menu.xml
@@ -11,7 +11,7 @@
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.canteen.app.LoginActivity">
+    tools:context="com.canteen.app.activity.common.LoginActivity">
 
     <!-- Login progress -->
