+++ app/src/main/res/layout/activity_food_details.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.canteen.app.activity.food.details.FoodDetailsActivity">
+    tools:context=".activity.client.details.FoodDetailsActivity">
 
     <TextView
         android:id="@+id/food_description_text_view"
