+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/ic_main_activity_background"
-    tools:context="com.canteen.app.MainActivity">
+    tools:context="com.canteen.app.activity.common.MainActivity">
 
     <Button
         android:id="@+id/login_button"
