+++ app/src/main/res/layout/activity_menu_details.xml
@@ -5,6 +5,6 @@
     android:id="@+id/foods_recycler_view"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.canteen.app.activity.menu.details.MenuDetailsActivity">
+    tools:context=".activity.client.menu.details.MenuDetailsActivity">
 
 </android.support.v7.widget.RecyclerView>
