+++ app/src/main/res/layout/activity_order_cart.xml
@@ -86,7 +86,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/appbar"
-        tools:context="com.canteen.app.activity.cart.OrderCartActivity"
+        tools:context=".activity.client.cart.OrderCartActivity"
         tools:layout_conversion_absoluteHeight="511dp"
         tools:layout_conversion_absoluteWidth="384dp">
